refactor: extract duplicated SQL filter logic into models::query helper
