+++ src/models/config.rs
@@ -23,27 +23,14 @@ pub(crate) async fn find(pool: &SqlitePool, id: String) -> Result<Option<Config>
 const ALLOWED_FILTER_COLUMNS: &[&str] = &["namespace"];
 
 pub(crate) async fn find_all(pool: &SqlitePool, filters: HashMap<String, Vec<String>>) -> Result<Vec<Config>, sqlx::Error> {
-    let mut query = String::from("SELECT id, created_at, updated_at, namespace, name, data, labels FROM config");
-    let mut all_values: Vec<String> = Vec::new();
-
-    if !filters.is_empty() {
-        let conditions: Vec<String> = filters
-            .iter()
-            .filter(|(k, v)| !v.is_empty() && ALLOWED_FILTER_COLUMNS.contains(&k.as_str()))
-            .map(|(column, values)| {
-                let placeholders = values.iter().map(|_| "?").collect::<Vec<_>>().join(",");
-                all_values.extend(values.clone());
-                format!("{} IN({})", column, placeholders)
-            })
-            .collect();
-
-        if !conditions.is_empty() {
-            query += &format!(" WHERE {}", conditions.join(" AND "));
-        }
-    }
+    let (query, values) = crate::models::query::build_filtered_query(
+        "SELECT id, created_at, updated_at, namespace, name, data, labels FROM config",
+        &filters,
+        ALLOWED_FILTER_COLUMNS,
+    );
 
     let mut q = sqlx::query_as::<_, Config>(&query);
-    for val in &all_values {
+    for val in &values {
         q = q.bind(val);
     }
